build: Add Dockerfile for building a custom ODR image.
The Dockerfile added in this commit can be used to add the Nomad Pack binary, from Nomad Pack tech preview version 3, to a custom Waypoint ODR image.

Future plans for this Dockerfile include automatically building and pushing this to a registry using Waypoint in CI, and templating the base image.