bugfix: Fix multiple issues in the ODR Dockerfile.
The Nomad Pack binary is now installed at /kaniko/nomad-pack.

The plugin binary is no longer copied from a zipped file. The raw binary is copied from the builder image.

The build step runs with CGO disabled. Prior to this, the plugin binary was not executable on the ODR image -  was returned.

Build only the linux_amd64 flavor of the plugin binary during the ODR Docker build to reduce build time.

Lastly, copy the Git CLI after installing to a plain alpine image. apk is not available for use in the ODR image, so we steal it this way.