bugfix: Base ODR image off of custom ODR base.
Copying the Git CLI from Alpine, where Git was being installed with apk, does not result in a runnable Git CLI in the ODR image. Because of this, I found it necessary to build the hashicorp/waypoint-odr image from source, while copying in all necessary components of Kaniko to the image into the base Alpine image. This enables the use of apk to install Git, which is required by Nomad Pack.